class and inherit example updates

+++ JavaProgramming/src/chap98_homework_nc230320/ArrayUtility.java
@@ -6,7 +6,7 @@ public class ArrayUtility {
 		double[] convertArr = new double[source.length];
 
 		for(int i =0;i<source.length;i++) {
-			convertArr[i] = (int) source[i];
+			convertArr[i] = (double) source[i];
 //			System.out.println(convertArr[i]);
 
 		}
